+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [0.2.20260209](https://github.com/Qiskit/ibm-quantum-schemas/tree/0.2.20260209) - 2026-02-09
+
+### Added
+
+- `executor` model version `0.2`
+- `noise-learner-v3` model version `0.2`
+
+### Removed
+
+- `executor` model version `0.2_dev`
+- `noise-learner-v3` model version `0.2_dev`
+
 ## [0.1.20260129](https://github.com/Qiskit/ibm-quantum-schemas/tree/0.1.20260129) - 2026-01-29
 
 ### Added